Change svelte to be the index page
Change former index page to be 'vanilla'
