+++ src/schema/simple-comment-openapi3.json
@@ -947,7 +947,7 @@
     "/auth": {
       "post": {
         "summary": "Accept a user name and password, return authentication token",
-        "description": "The authentication token is set via the Set-Cookie header with a 'simple_comment_token' key or as an Authentication header with Bearer scheme, depending on Simple Comment configuration. If the authId is a temporary user, link the temporary user to the authenticated user",
+        "description": "The authentication token is set via the Set-Cookie header with a 'simple_comment_token' key or as an Authentication header with Bearer scheme, depending on Simple Comment configuration. If the authId is a temporary user, link the temporary user to the authenticated user. Special case: If a user claiming to be the 'Big Moderator' is not found in the database but provides the correct password, they are authenticated and an auth token is returned, but no user object is created in the database.",
         "requestBody": {
           "required": true,
           "content": {
