+++ src/static/index.html
@@ -23,8 +23,68 @@
       content="Simple Comment logo: thought bubble with a slight smile"
     />
     <meta name="twitter:site" content="@rendall" />
-    <link rel="stylesheet" href="./css/index.css" />
+    <link
+      rel="stylesheet"
+      type="text/css"
+      href="./css/simple-comment-style.css"
+    />
   </head>
+  <style>
+    body {
+      -moz-osx-font-smoothing: grayscale;
+      -webkit-font-smoothing: antialiased;
+      background-color: #f8fcff;
+      color: #33475c;
+      font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Helvetica,
+        Arial, sans-serif;
+      font-size: 1.125rem;
+      line-height: 1.61111em;
+      margin-bottom: 1em;
+      margin-top: 1em;
+      margin: 0 auto;
+      padding: 0 1rem;
+    }
+
+    a:hover {
+      color: #007b8c;
+      text-decoration: underline;
+    }
+
+    a {
+      color: #3287e6;
+      font-weight: 600;
+      text-decoration: none;
+    }
+
+    header {
+      display: flex;
+      margin: 0 auto;
+      max-width: 59rem;
+      min-height: 10rem;
+      align-items: center;
+    }
+
+    .simple-comment-logo {
+      max-width: 9.1rem;
+      min-width: 9.1rem;
+      fill: #3287e6;
+    }
+
+    main {
+      max-width: 38rem;
+      margin: 0 auto;
+    }
+
+    p > span {
+      color: #3287e6;
+      font-size: 1.2rem;
+      font-weight: bold;
+    }
+
+    section {
+      margin-bottom: 3em;
+    }
+  </style>
 
   <body>
     <header>
@@ -48,71 +108,66 @@
         </g>
       </svg>
 
-      <h1>Simple Comment</h1>
+      <h1 class="simple-comment-logotype">Simple Comment</h1>
     </header>
     <main>
       <section>
         <p>
-          <i>Simple Comment</i> is a commenting platform for any website. Free,
-          flexible, open-source, private, and scalable.
+          <i>Lorem ipsum</i> dolor sit amet, consectetur adipiscing elit. Ut eu
+          velit pulvinar, commodo risus vel, finibus quam. Curabitur iaculis
+          aliquet urna nec semper.
         </p>
         <p>
-          <span>Free:</span>
-          <i>Simple Comment</i> takes advantage of free-tier offerings from
-          cloud service providers such as Netlify and MongoDB Atlas to remain
-          no-cost indefinitely.
+          <span>Nam tempor:</span>
+          <i>Mauris luctus</i> aliquam eros quis posuere. Praesent vitae libero
+          augue. Duis a laoreet turpis. Sed dapibus eu orci vel convallis.
         </p>
         <ul>
           <li>
-            <a href="https://www.netlify.com/pricing/">Netlify Starter</a>:
-            125,000 API calls/month with 100GB Bandwidth
+            <a href="#">Nunc a varius sapien</a>: Donec ut enim tempor,
+            consectetur risus nec, bibendum dui.
           </li>
           <li>
-            <a
-              href="https://www.mongodb.com/docs/atlas/reference/free-shared-limitations/"
-              >MongoDB shared M0 Cluster</a
-            >: "Free forever" with 512 MB of data
+            <a href="#">Donec volutpat enim tempus</a>: erat tincidunt, eu
+            vestibulum orci ornare. In sed imperdiet ligula.
           </li>
         </ul>
         <p>
-          <span>Flexible:</span> <i>Simple Comment</i> is designed to easily
-          swap in other services or vendors as needed.
+          <span>Vivamus eleifend:</span> Nunc imperdiet mattis ipsum, ac posuere
+          mauris rutrum in. Duis non mi sit amet quam lacinia molestie.
         </p>
         <p>
-          <span>Open Source:</span> The backend source code can be freely used
-          and modified under the
-          <a
-            href="https://github.com/rendall/simple-comment/blob/master/LICENSE"
-            >GNU AGPLv3 open source license</a
-          >. The running API can be accessed with any client under any license.
+          <span>Duis finibus:</span> Mauris ut congue turpis. Cras lectus
+          ligula, maximus ut tristique at, consectetur nec libero.
         </p>
         <p>
-          <span>Private:</span> <i>Simple Comment</i> tracks no one and hoards
-          no data.
+          <span>Donec aliquet:</span> Aenean a elit at diam interdum ultricies.
         </p>
         <p>
-          <span>Scalable:</span> Each service provider offers upgrades of their
-          free tiers.
+          <span>Curabitur imperdiet:</span> magna vitae congue sollicitudin,
+          quam dolor eleifend leo, eget eleifend turpis dolor eget mi.
         </p>
         <p>
           <span>Affordable Pricing:</span>
-          <a href="./pricing.html">Choose from our range of pricing tiers</a>,
-          including Free, Basic, Pro, and Enterprise options, to find the
-          perfect plan for your needs. Whether you're looking for a no-cost
-          solution or a customized package, Simple Comment has you covered.
+          <a href="#">Nunc feugiat dui ut nulla posuere</a>, in semper ante
+          malesuada. Phasellus consectetur ligula vitae metus imperdiet
+          convallis.
         </p>
         <p>
-          <span>Getting Started:</span> Try the demo below,
-          <a href="./pricing.html">explore our pricing options</a> and
-          <a href="https://github.com/rendall/simple-comment#readme"
-            >read the documentation</a
-          >.
+          <span>Getting Started:</span> Vivamus dapibus est sed ipsum elementum
+          cursus. In tempus pharetra varius.
+          <a href="#">Nunc ornare hendrerit lacus</a> and
+          <a href="#">Proin tempus arcu a orci semper</a>.
         </p>
       </section>
-      <section class="comment-display">
-        <div id="simple-comment-display"></div>
-      </section>
+
+      <section id="simple-comment-display"></section>
     </main>
     <script type="module" src="./js/simple-comment.js" defer></script>
+    <script type="module" defer>
+      var params = new URLSearchParams(window.location.search)
+      var topicId = params.get("topicId") || "http-localhost-7070"
+      setSimpleCommentDiscussion(topicId)
+    </script>
   </body>
 </html>
